Add Fullscreen, properly fix camera

+++ Engine/Camera.h
@@ -7,15 +7,15 @@ class Camera : public Object
 
   private:
     Maths::Matrix4 projection;
-    int xangle;
-    int yangle;
+    float xangle;
+    float yangle;
 
   public:
     Camera(float fov, float ar, float near, float far);
     Camera(float left, float right, float top, float bottom, float near, float far);
     ~Camera();
     const Maths::Matrix4 &Projection();
-    void turn(int x, int y);
+    void turn(float x, float y);
     void lookAt(const Maths::Vector3 &v);
 
   private:
